Move e2e tests out of beta dir
