+++ test/e2e/contract-test/contract.js
