+++ test/e2e/metamask-ui.spec.js
@@ -9,7 +9,7 @@ const {
   installWebExt,
   getExtensionIdChrome,
   getExtensionIdFirefox,
-} = require('../func')
+} = require('./func')
 const {
   assertElementNotPresent,
   checkBrowserForConsoleErrors,
