+++ test/integration/lib/send-new-ui.js
@@ -4,7 +4,7 @@ const {
   queryAsync,
   findAsync,
 } = require('../../lib/util')
-const fetchMockResponses = require('../../e2e/beta/fetch-mocks.js')
+const fetchMockResponses = require('../../e2e/fetch-mocks.js')
 
 QUnit.module('new ui send flow')
